Add generic extension mechanism
By using the shell mechanism we can actually provide an more generic
extension call-back mechanism. We can identify and match on the
control word, which gives up to 256 different call-backs.

The mechanism seems quite powerful, allowing generic
notifications (somewhat resembling poll vectors), as well as
extensions to commands, i.e., CAT could invoke a specific generic
extension if the catalog number is not defined and the extension on
the shell could act on it, providing further functionality.