Basic keyboard handling working
The system digit entry flag is also used by application shells as
alarms can terminate digit entry.
Thus, digit entry flag may be set for application shells and we should
consult the shell stack in this state.
Made various corrections to custom application keyboards to make it
work (at least a bit) for Ladybug.
Fix the system/application/extension nibble handling which was not
done in an entirely consistent way.