Correct shell disabling and reclaiming
This code was more subtle that I thought in the last change. Both the
disable and reclaim routines share this code. As the reclaim code has
the page number reset, we need to search for the shell testing all
nibbles but the page number. Make this behavior more obvious in the
comments. Also the idea is that once found, we use the potentially new
page as the page for the shell. That way it is possible to turn the
calculator off, move the module to another port and the reclaim code
will properly mark the shell descriptor to belong to the new page.