Clear digit entry on key fall-back
If a shell does not define a key, it is assumed that the default
function will end digit entry. This makes it work well for keys such
as XEQ and ASN, but it also means that PRGM, SHIFT and USER need to be
defined by the application keyboard. This is probably not such a big
issue as such mode changing keys are probably not going to have an
override.

For XEQ and ASN, special functionality is more natural, to extend
their behavior. It is therefore a good idea for an application shell
not to define them and rely on some system extension to implement such
behavior.