Proper handling of back-arrow key
Introduce an internal Flag_DisplayOverride that works together with
the system message flag. The display override flag tells whether the
display is showing some custom view, rather than a system message.
We use the message flag to prevent normal system display of the X
register. However, we still need to support that a system message is
shown and that the back-arrow key is used to remove it. This case is
now properly handled in a generalized way.

The generic code now handles back-arrow for everything except digit
entry and custom clear-X functionality, which is what the shell
specific keyboard code needs to deal with. This makes it a lot simpler
for the shell code that can focus on only its own stuff, without
any system related behavior of back-arrow.

The two flags are interpreted as follows:

Message   DisplayOverride
1         1         we are displaying an override
1         0         system is showing a message
0         x         nothing is done, do display override if needed