Preliminary fix for catalog key input
The built in catalogs 1-3 act on actual keyboard input and undefined
keys will abort the catalog and perform the keyed action. This gets a
bit in the way with redefined keyboards. What is interesting is that
the actual function bound to a key as what determine this. Thus, the
ENTER key may be redefined and not perform the 'enter' action in
catalog 2. The ENTER function may be assigned to another key and that
key will perform the 'enter' action in user mode.

This requires that we need to anticipate what should happen depending
on the system behavior and this is not entirely trivial. For now, we
settle for assuming that these system keys are not redefined (which is
likely), but also that such operation has not been assigned to another
key (like BST to the TAN key). This is somewhat more likely to happen.

In any case, this is a temporary work-around that makes it work in
an ordinary setup. Later, it should probably have a more elaborate
code to deal with all corner cases.