Rework shell find logic
It was flawed in a couple of ways. It would not find system shells
before application shells in the stack. In reality, it makes sense
that whatever is on top of the stack shadows the rest, no matter of
its status.

The root problem was that the shell stack scanner did not properly
separate shells from extension points. This is now done by adding the
topExtension routine (instead of topAppShell) and have the logic so
that either class is (only) returned. When searching for shell,
applications are always accepted, but only the first one.

Also cleaned up some comments.