Make extension handler scan a list of handlers
Each shell costs half a RAM register and RAM is a scarce
resource. Thus, it makes sense to define as few extension points as
possible and it turns out that a linear list of handlers allow for a
module to define a single extension shell, but have multiple handlers
in it.