Define shell kind bits differently
Create a simple way of marking a shell as a transient application
shell. Thus, we change the bits so that we can test them by bit left
shift (c=c+c). Then the transient application can both set the
application bit and its own transient bit, making it easy to test both
system or any applications as before.

Also, there is not need to fetch the kind word from the descriptor as
it is already loaded into the xs field of a right aligned shell
descriptor, which in part is the point to have it there, to make it
easy to categorize the different variants.