Clear the transient scratch pad during power up
Transient applications are not meant to survive power downs. In theory
they could, but it is judged that powering off is a good way to clear
such behavior as that is also done with built in catalogs, and these
typically belongs to that category.

In case the system went to power off with a transient scratch area
allocated, it is a good idea to get rid of it when coming back on as
it has no current use. It can easily be created again as needed.