Organize shell stack based on shell kind
Always ensure that a transient application goes on top of the shell
stack. Immediately after it comes the ordinary application shells and
then everything else.

The reason for this is to ensure correct behavior according to the
conceptional model. There are in general three rules.

1. A transient application always has precedence over any application.
   It is meant for temporary modes, such as catalog that can have the
   ability to auto-terminate on an undefined key.
2. The top-most ordinary application is then supposed to take key
   input (shadowing any other application).
3. Finally, system shells are consulted in the order they appear for
   handling keys not defined by the top shell.

Applications are separated from system shells to ensure that an
application is always consulted first. Keeping the order on the stack
makes the traversal code simpler (and faster) and using the shell
order on the stack is done much more frequently compared to making
changes to it (typically done on activation and during power on).