Correct problem with orphaned shells being used
At power on it would release the shells and allow the owners to
reclaim them. However, it would invoke the display routine before it
actually invalidated any purged orphaned shell. If such orphaned shell
was at the top of the shell stack it would point to a random location
in page 1, causing it to jump into the middle of nowhere causing
severe problems.