Add anonymous XKD keys for sparse keyboards
These are intended for catalog style functionality. There are no
preview on certain keys which act immediately instead. Provide a
special XKD feature so that instead of actual XROM entries, one can
have anonymous pointers to key handlers that act on them. It avoids
polluting the FAT and makes it easier.