Add more outputs from 'argument'
Also pass out the argument result in A.X and C.X, to make it easier to
use. It is not so dangerous to have the output in a couple of
registers as this entry point is intended to be first thing that
happens in instructions that make use of the feature.