Improve buffer layout
Allow the hosted buffer area to be larger than 15. It is better to
limit the number of secondary KARs to 15 (allowing up to 30 such
assignments, which is a lot).

Make all sub-fields in the buffer being sizes as it makes it easier to
maintain that way, we calculate offsets when used. This should make it
a lot less error-prone.

This makes is easier to step secondary KARs and we do not need to
provide a routine for that anymore.