Initial support for custom partial key sequences
Allow XROM instructions to have custom partial key sequence
handling. Such XROMs are marked using 'gosub partialKey' and its name
header puts up the initial prompt underscores in the usual way.

When such XROM instruction has started, it is intercepted and partial
key sequence prompt handling is redirected to the XROM itself.

This way an XROM can act as a prompting instruction properly, putting
up initial underscore(s). Due to a minor bug in the HP-41 mainframe,
NEXT (which puts up the underscore(s)) goes to WKUP10 which bypasses
the 0x4000 vector. As a result, the actual intercept takes place when
the first key is pressed. Doing it immediately would require that the
I/O poll vector is used.

The advantage of doing this way is that we get proper XKD behavior on
key down for such instructions, getting around the XKD/XROM bug.
It also saves code space as we get the prompt name set up by just
using the appropriate bits in the name header.