Allow complete range for secondary assignments
There is not need to have the F0 prefix inside the buffer. Omitting it
gives another nibble for the key assignments which simplifies the code
as we do not need to pack the XROM Id and function number together. It
also gives full range (0-4095) secondary function number assignments
to keys.