Clean-up handling of built-in functions
Keyboard definition using built-in functions was kind of done in two
places, though only one was actually used. Removed the duplicate code
and beefed up the one that was used to work the same way.

This makes use of adjust function number for built-ins as 000 is used
for pass-through. Normally you will access any original function by
using pass-through rather than explicitly inserting a built-in on the
standard keyboard. There are probably use cases where putting a
different built-in on the standard keyboard and doing it this way
makes it possible to explicitly have standard CAT on the keyboard (as
it has number 000 in a keyboard definition!)