Add a timer module
Allow the interval timer of the Time module to be used by application
shells and partial key sequences.

Clean up the 169 cold start constant related to MEMCHK. This was a
leftover from the Extended Memory cache mechanism that used this
constants to speed up access, but this was deemed too dangerous and
abandoned. However, some old code remained that is not cleaned up.

Doing it for partial key sequences is currently untested.