Notify when shell stack is altered
Remove the activation and deactivation vectors as it is too tricky to
implement them. Use the notification mechanism instead which is more
flexible (something is changed with the shell stack), though it puts a
bit more burden on a normal shell that wants to use the timer when
being active.