Fix display of secondaries in program memory
Do not emit any line number until we know for sure that we can
actually find the secondary function. Previously it would output the
line number and then look up the secondary function. If the index used
with the XROM prefix in program memory was out of range, then we do
not have any secondary function and bail out. However, we have already
prepared the display with a line number, so it ended like that,
without any instruction display and right justified.

A secondary can be out of range if the text literal following the XROM
prefix (for secondary in program memory) was bogus, or the secondary
table was update to remove some functions at the end.

Basically, we reorder the work so that we look up the line number
early as it may trash a lot of registers, but we now postpone writing
to the display until we know that we can locate the secondary
function. If we cannot, then we just bail out and use the
existing (correct for the situation) program memory display.