Better support for semi-merged secondary functions
The dual execution of the function caused problems for secondary
functions as the execution of such is a bit special (it uses A001 aka
XROM 0,1 which is invalid, combined with the XADR in M) and needed
some work outside program mode. Prompt handling is most often handled
by the main operating system and it has no idea about secondary
function execution. To make it work we use the XROM prefix
instruction (same as is used inside programs) and need the function
number tucked away (in register 9/Q) so that it can be found by the
XROM prefix instruction. The runSecondary routine now takes care of
the actual execution outside program mode for semi-merged
instructions.