Fix comment
Take a buffer number in the exponent field. The code was written that
way and the instruction set makes it easy with the LDI instruction to
have it like that. Just using the [1:0] could be fixed with a clear
exponent field, but it wastes and instruction and seems like no
point.