Do not use N register in chkbufHosted
Instead use G as it can be good to leave the N register alone.

This was mainly done for 2D6 routine that calls this routine when it
wants to save the updated seed. It has both a result and a new seed to
store so it needs two registers to hold the two values (N and M).
Here the N register was used for storing the buffer number, but G is
more seldom used and is a suitable alternative.