Remove unclaimed hosted buffers at power on
Also fix a problem in that it was a bit too aggressive to remove the
scratch area every time it was going to light sleep without any
special message in the display. The code that uses the scratch area
should remove it when done, and it may be a good idea to allow it to
stay around longer than that.