Remove no-applications flag
We are running out of flags need to take this flag bit for more
important duties. This flag is also also dubious and the ability to
push a standard shell is probably better done by having an ordinary
shell descriptor which can be ordered in the same way as other shell
descriptors.