Reset the secondary proxy flag after use
This flag is intended to be used once and was not properly reset which
could potentially lead to problems. There are actually several things
that need to be fulfilled for it to cause problems, but it is not good
to leave stale flags behind that is not telling the truth about the
current state.