Keep primary bank in XASRCH
Do not switch to the bank a secondary function reside in, instead
return with primary bank enabled. All existing consumers expect to
have the primary bank enabled. For assignment we do not care about the
bank and when executing (XEQ) we essentially does the same as in
mainframe by looking up the function by its identity for execution
looks at the secondary FAT header first.