Correct wake up from deep sleep
There was a problem in that it would look for the system buffer and
trust its contents at power on, before doing any memory integrity
checks. Also perform the check for master clear before looking for the
system buffer. This ensures that the calculator does not get stuck
reading corrupt data and allows for a master clear to get out in a
state that is incorrect, but not seen as corrupt by the rather
simplistic sanity checks.