Make OS4 banked
Open up a second bank to make things fit. The alternative is to lift
out routines to the page relocatable modules, but it is judged better
to have fixed addressing in page 4 for routines that may be useful
from multiple modules. The slight disadvantage is that not all memory
systems on the market supports it at the moment, but hopefully that
can be addressed over time.