Clean up output from argument
It does not need to return it in so many places. ST is a good choice
as mainframe expects it there for its support routines. Ladybug likes
it in G as it often mess up registers a lot in preparation and G tends
to be a rather safe place to put it.