Correct clock display termination
The clock display would not properly detect termination on key pressed
when a shell that override the display was active. This is because the
clock function uses a simplistic way of detecting if a key has been
pressed. It inspects the message, digit and partial key
flags. Application display override may set the message flags to
reduce flicker it will fool the clock display that no key has been
pressed, which means it would not terminate as it should.

As it is not possible to fix the time module, the problem is worked
around by sensing a key press with an active application, in that case
the clock display is now terminated (if there is a time module
present).