Allow duals to allow stack input independently
It makes sense to allow a dual semi-merged function to request
arguments where one allows stack and not the other.

Previously the two upper bits could be used from the control word to
control prompt behavior, but in reality we only want to express
whether it allows stack operands or not, thus a single but is enough.

The freed bit now change meaning to be the stack allow bit for a dual
operand function.

Add some constants in the header file to make it easier to build the
control word that requests stack addressing or not.