Add displayingMessage routine
This is intended to help writing shell style backarrow key
functions. When a message is shown it should clear the message, not
clear the X register. As the system message flag is overloaded when
using a shell and the internal display override flag is hidden, this
routine exposes the message display status properly.