Put up SS0 before calling a function
Make the CPU state on function entry more similar to what it is for
primary functions. The book 'MCODE for beginners' lists several things
about the initial state, some of them are a bit too specific. However,
SS0 is sometimes useful and somewhat relevant so we now put it up
before calling the function.

On one hand secondary functions do not have to have exactly the same
initial state as a primary functions. On the other hand, it does not
hurt to follow it a bit where relevant and functions may be moved
between being secondary and primary. This makes it easier to do so.