Do not expose 'secondaryAssignment'
This routine is for looking up an assignment on a given key. The
problem with exposing it is that 'testAssignBit' also should be
exposed as they are used together. The output from 'testAssignBit' is
what 'secondaryAssignment' expects and 'testAssignBit' has multiple
exit points, making it tricky to expose when it is in bank 2.
As these routines are probably not that interesting for outside
module, I decided to not have them exposed for now.