Correct direct execution of lonely prefix XROM
When executing a secondary semi-merged function from the keyboard in
run-mode its associated prefix XROM is used and the adjusted secondary
function identity is stored in REGN9[1:0]. This allows mainframe to
execute the secondary function when the argument is done.
So far so good, but if a secondary function is executed by hand,
i.e. it has been assigned to a key or is executed by its XROM
number (using XEQ' in Boost), there is no adjusted secondary identity
available, as we are not executing any secondary function.
In this case runSecondary would pick whatever was in REGN9[1:0] and
assume that was an adjusted secondary identity, but it was just
garbage.

Now we tag REG9 that there is a valid adjusted secondary function
identity in REG9[1:0] and also keep a counter in REGN9 for how many
calls to runSecondary it should stay valid for, which is the number of
postfix arguments! After that it becomes invalid and we store
something else in REGN9 to prevent it from matching any separate
invocation of a prefix XROM.

This is not entirely bullet proof as we share REGN9 with mainframe and
other things, but it should provide a good amount of safety. In any
case, executing prefix XROMs outside program mode by hand is not
really meaningful in the first place, but now we provide a fairly
solid guard against doing it by accident and having a random secondary
function potentially being executed.