Major rework on the catalog support code
Support banked catalogs in application modules.

Provide a set of support routines that are mostly called as a result
of pressing keys when the catalog is stopped and is using a transient
application for input.

The transient application is owned by the application module and the
keyboard handlers are in its bank 1. These are in most cases simple
stubs that calls one of the new points. The catalog support code
figures out when to step and does it using entry points in the catalog
descriptor provided by the application module.

We can step both forward and backward. Support for doing other
actions (like remove an entry) and setting a new position from the
application module is provided for.

Also update the reference part of the documentation.