Add jumpC7 and unpack7
It is fairly cheap to add these (two words) and it not as easy to do
it after the initial release. This gives one more entry than is
actually used today, but they are general routines and may be useful
when OS4 is expanded in the future. It also gives 8 such entries (0-7)
which is more even in the binary world. Put together, it makes a lot
of sense to add them.