Fix problem with added catalog and clock display
When having the clock display updating, pressing CAT 07 caused the
default catalog 3 to be displayed.

This was caused by a combination of things. The clock display actually
turns the HP-41 off, so we will enter reclaim mode. During reclaim
mode notification messages are not sent, and the catalog mechanism
uses such messages. Reclaim mode was a bit of an add-on that was meant
to be fixed when going to light sleep, however this did not happen in
this case.

Made it more robust by copying more of the power on code from mainframe
to OS4 so that we can remove orphaned shells during the power on
handling. This makes the code a lot simpler at the expense of some
minor ROM space. As an extra bonus it freed a flag bit, yay!