Gracefully handle when there is not system buffer
The start-up code had an issue if the system was started without any
module creating the system buffer. The problem was that
disableOrphanShells should jump back as it uses too much stack, but
when no system buffer existed it would return to 0000, causing the
calculator to fail to turn on properly.

Having Boost inserted when starting would not expose this problem as
it always create its system shell on power on.