Correct ensureHPIL
It would most likely fail even when an HP-IL module was attached due
to that the pointer was not set when building the 7000 address.

Also clean up the documentation related to this and make ensureDrive
first check for HP-IL, then if there is a mass storage drive present.
This adds some safety for modules misusing page 7 and provides a
slightly more accurate error response as it even can tell you forgot
insert the HP-IL module in the first place!