Fix executing secondary functions with printer
Calling the printer to print the function about to be executed did not
work well at all. First it clobbered the M register where the
information about the secondary function is held, second it printed
the function based on a function code and that only works for primary
functions.

For now I just disabled printing the function name.