Fix secondary semi-merged functions
These would give NONEXISTENT when running from a program if the
argument was the default one. There were actually two places it had
problems, first it would not accept that the Text literal count could
be one less (which it is in this case), then when actually picking up
the default argument the test for one less was reversed.