Do not reset pause flag when pressing keys
The pause flag is to be reset when the pause times out or if the R/S
key is pressed. It should not be reset just because we perform some
calculations or end digit entry, as it is perfectly OK to press a
sequence of keys to perform calculations while pausing.