Add pausingReset function
This is intended for STOP functions that terminate program
execution. Such function wants to know if we are currently pausing (in
which case it should stop) or not. This is because the program
execution flag (13) is cleared while pausing and we just pretend to be
running by showing the PRGM annunciator.