Provide a hide auto-assignment flag
Use the available system buffer flag as a flag that disables
assignments on the top keys and obey RPN label auto-assignments.

This flag survives cycling the power, the rest of the flags are of
more transient nature and are always reset.

Also fix a bug with auto-assignments as it did not entirely work
before. It would clobber the shell scan state and as a result ended up
in an infinite loop when given a sparse keyboard (like the system
shell in Boost).